Update patch set 1
Patch Set 1: Fails

(1 inline comment)

Thanks for the contribution, that was fast!

In addition to a few nitpicks with the code style, it'd be good to format the commit message a bit differently: typically, it's best to start with a line of 65 characters or less, followed by a blank line, followed by lines of 70 characters of less.

Besides the small style issues, the code merges and builds fine.

Patch-set: 1