Honor exported=false attribute for providers.
A provider defined in an AndroidManifest.xml file may set its `exported'
attribute to false (the default value is true), indicating that only
activities running in the same process as the provider may access it
(except root and the system process, which are always granted access).
This patch fixes an error which caused the exported attribute to be
ignored, thus granting any process access.

Change-Id: I5e12875893c785629a2f57362a190d340926f4f6