Update patch set 2
Patch Set 2:

We just want to let you know that we are working on this.

From Author:

* sizeof (line 64)
Done. (The char* pointers have been turned into char[] arrays.)

* session lookup lock
Done.

* SEEK_END
The end of the file coincides with the end of the encrypted data, so no offset relative to the end is necessary. The return value, however, needs to be adjusted, which is done a bit further down. A comment has been added to make this clearer.

* read return value check (line 378)
Unless the size of the embedded content file happens to be an exact multiple of the buffer size, the next-to-last read() will return 0 < numBytesRead < sizeof pData->buffer. The very last read() returns zero bytes and breaks the loop.

* off_t
A valid point that merits investigation, but the Forward Lock plugin is just implementing a DRM Framework interface here, so it can't be unilaterally changed.

Other items:

> can omit the "? 1 : 0" as "<" returns an int of the proper value [FwdLockFile.c(217)]
That would make the expression Boolean and give it type (int), and the result would have to be typecast back to (unsigned char) to avoid compiler warnings.

> == 0 could be !memcmp (optional) [FwdLockFile.c(386)]
Our in-house style is to not use the Boolean operators on non-Boolean operands, so unless Android guidelines mandate otherwise we'd prefer to keep it as is.

Patch-set: 2