Update patch set 2
Patch Set 2: Verified

While it's hard to check the style in detail given the size of the change, a random sampling doesn't seem to find problems, and I trust that you have a history of closely following the official style guide.

The code merges and builds fine both in the open-source master tree and in Google's internal master tree.

As a note, if you'd like to get rid of .. in makefiles (which isn't strictly mandatory as those point to true source files, not to generated files), a good trick is to move all the makefile logic up in the tree from where it can "see" all the necessary files. It doesn't always solve everything, but it can help.

Patch-set: 2
Label: Code-Review=0
Label: Verified=+1