Update patch set 1
Patch Set 1:

It's my understanding that CDMA was largely copy/pasted from GSM and the "base class" isn't much of an actual base class.  So the goal is to start cleaning that up and pushing stuff down?

Seems a little unfair to push that burden onto contributors.  In cases where it's too hairy, would it also be acceptable to simply mirror the change in the CDMA version?

Patch-set: 1