Correctly calculate the FastScroller's scrollTo position when using ExpandableListView.
The current code calculates the target as a flat list position, but then
incorrectly uses it as a group index by passing it to
getPackedPositionForGroup(). As a result, if there are expanded items,
we will potentially scroll to the bottom of the list much too early.

This patch assumes that the desired behavior is having the fast scroll
feature target the individual groups. The alternative would be to target
each individual flat list item while scrolling.

Change-Id: Ie62f781c26edcf3073d7818cd37eb9fd4a8aa14d