SurfaceFlinger: Don't update overlays while frozen
Surface flinger can be frozen in which case it won't draw or post
anything to the display. This mode is used for example when changing
orientation to show all the changes at once. Currently overlay changes
are still sent to the overlay device when surface flinger is frozen
which causes the overlays to be updated significantly before the UI in
for example orientation changes. By deferring the overlay changes till
surface flinger is unfrozen a better result is achieved.

Change-Id: I9e8cd9c3d0043a203a27b3539481c513f7a1739f
Signed-off-by: Johan Mossberg <johan.xx.mossberg@stericsson.com>
Signed-off-by: Christian Bejram <christian.bejram@stericsson.com>