Update patch set 1
Patch Set 1:

vairav,

We should probably make the class static and all of the fields final, that would be good coding practice and an oversight on our part. But I don't understand why any of those things would make it collectable.  So could you provide a reference to a doc or explain why making mBinder final allows it to be collectable and if we don't its not. Note, its not that I don't believe you I'm just trying to enhance my understanding.

Thanks,

Wink

Patch-set: 1
Label: Code-Review=0