Update patch set 1
Patch Set 1: Do not submit

This check needs to be there because in some countries (Japan IIRC), the system in incapable of making calls until the SIM is unlocked.  If it's PUK locked, it'll never be able to make an emergency call in those places.

That being said, if you look at the method it calls to make the decision, you'll find the configuration variable it looks at to make the decision.  It can be overloaded in a specific build using overlays if you find a country where it should be enabled.

Patch-set: 1
Label: Code-Review=-2