Update patch set 1
Patch Set 1:

Conley: Is there still a reference to android.R.drawable.ic_dialog_alert in the code? If not it might not be needed in the internal tree. If there is, is it perhaps more like the references in the source for Bluetooth.apk where I'm not 100% sure how fix them?

Patch-set: 1