Update patch set 1
Patch Set 1:

A property would work well for us.

Some background on why we are crashing the system:
We tend to use full system dumps as the easiest way to get full information on a problem in a way that any  user can easily send to the development organisation. In debug variants, full system dumps are taken on kernel crashes, and stored on /sdcard. Via system properties, we can enable application crashes to cause full system dumps as well, if we are chasing a particularly difficult problem.
In the case of system server crashes, we are always restarting the entire system, even in user variants, to avoid state mismatch problems between app and modem, so in our case this was just a way of moving the kernel panic a little earlier, while it still contained useful information about the system server process.

Patch-set: 1
Reviewer: Gerrit User 1015615 <1015615@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=0