Update patch set 1
Patch Set 1: Code-Review+2

This is fine, but please be aware that code running into this is probably suffering from other race conditions.  It is really only safe to do view hierarchy calls on the same thread that the view hierarchy is running in.  Code that is calling this from other threads is going to have other race conditions -- for example it may get a Handler back but the view hierarchy is no longer attached to it.

Patch-set: 1
Reviewer: Gerrit User 1000559 <1000559@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=+2