Update patch set 1
Patch Set 1: Code-Review-1

I don't think this is the right solution.

First of all, if the issue is that the first task can be killed, we should just prevent it from being killed, rather than hiding it entirely.

However, I also disagree that the current behavior is undesirable -- I know a lot of people that use it to kill the current app if it's running into problems.

Finally, showing the first task is an integral part of the recents animation (ie when the current task shrinks down into Recents).

Patch-set: 1
Reviewer: Gerrit User 1022558 <1022558@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=-1