Update patch set 1
Patch Set 1:

Hmm, I can't make up my mind about this change. It makes things a bit clearer, but this isn't code Elliott or I actively maintain, so we can't really ensure that new methods will follow this convention.

How about we change / fix the definitions that are wrong (the "possible bugs" you mention) and put the rest of this change on ice for now ? (I'll try and get some consensus on this).

enh@ : What do you think ?

Patch-set: 1