Update patch set 1
Patch Set 1:

I agree with your sentiment that keeping changes to the minimum is a good thing. It bugs me though, that these values are being downcast to jint from 64 bit unsigned values. I'll confer with the owners and try to get them to change their code.

Thanks for being very thorough in your analysis of this code.

Patch-set: 1