Update patch set 1
Patch Set 1:

If there are errors in the framework's GL context then we probably do want to crash, as we are now in an unknown state. If an app is making GL calls with framework's GL context (something no longer possible in L), then that's an app bug.

Do you have an example of a situation where this is both not an app bug and something we can expect to gracefully recover from?

Patch-set: 1
Reviewer: Gerrit User 1010103 <1010103@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=0