Update patch set 3
Patch Set 3:

I don't understand how the fix corresponds to the problem description.
This callback code is executed in the app process, so if you kill the app it will not be executed.
The DEAD_OBJECT error means the mediaserver process died, not the client app.

Patch-set: 3
Reviewer: Gerrit User 1003259 <1003259@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=0