Enable notifications during an ongoing call
If a phone call is ongoing, the alert will be allowed but only
using a special incall notification tone and no vibration. The tone
will be played on the voice call audio stream.

Also enable customization of the relative volume of the in
call notification volume. This allows the level to be changed for
different operators using overlays.

The requirement behind this is to allow audible SMS/MMS notifications
during phone calls, something that is requested by a number of
European operators.

Bug: 28688969

Change-Id: Ia8ce5c784a847e6759b55e6a78ac2100db47df19