Update patch set 7
Patch Set 7:

Hi again,

after the comment from Robert in buganizer I dug out some older devices (I even got my hands on a G1!) and tested a little with the following result:

 1.6   - Not possible to enter more than 8 digits into the PIN entry field
 2.3   - Not possible to enter more than 8 digits into the PIN entry field
 4.3   - No error message, it attempts unlock and then just clears the input
 4.4   - No error message, it attempts unlock and then just clears the input
 5.1.1 - Error message reported "SIM PIN operation failed!"
 6.0   - Error message reported "SIM PIN operation failed!"
 7.0   - Error message reported "SIM PIN operation failed!"

I've dug a little deeper into our internal error report and it seems that this was an internal regression only reported on our version of Jelly Bean, so apologies for the misinformation.

So this change is really just a usability issue.  If there are concern about other OEMs not sticking to the GSM specs of 8 digits and locking end users out, then I won't push for this.

Patch-set: 7
Reviewer: Gerrit User 1004882 <1004882@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=0