Fix unusual fade off animation when removing language
When change all menu item display property from SHOW_AS_ACTION_NEVER
to SHOW_AS_ACTION_ALWAYS, it will call ListPopupWindow.show()
finally. Since the menu item is changed to show as action, the popup
window height is changed to 0 and then the popup window will be
updated to the wrong position with animation. So it would be better
to skip operation at that time.

Bug: 35367495

Change-Id: I78205553ba558d6d66ea98238111788ecbb44912