Do not expose unused resources in overlay packages
Teach ResTable::getResourceIdentifer to ignore overlay packages. When a
client asks for a resource identifier in a package, the lookup should be
made only in the context of that package, i.e. the package and any
related split packages, but excluding any overlay packages.

This prevents the following scenario:

  - target package:
    - drawable/a.png

  - overlay package:
    - drawable/a.png
    - drawable/b.png

ResTable::getResourceIdentifer would return a resource ID even for
drawable/b. The resource ID makes no sense in the context of the target
package, and the actual numerical value is only valid within the context
of the overlay package.

Co-authored-by: Olof Axelsson <olof2.axelsson@sonymobile.com>
Co-authored-by: Ove Hanebring <ove.hanebring@sonymobile.com>
Test: run libandroidfw_tests
Change-Id: If7a9576f5db8555b0817f0bb6abe50990ca24e9c