Connectivity metrics: register every new network
This patch adds a non-Binder method to IpConnectivityMetrics for
registering new network, intended to be used by ConnectivityService.

This allows to track physical link layer for all connectivity events by
maintaining a map of network ids to link layers in
IpConnectivityMetrics.

Link layer inference is based on the interface name of LinkProperties.
IpConnectivityMetrics will also remember interface names which were not
recognized, so that post-inference can be done from serialized metrics
data.

Test: TODO
Bug: 34901696
Change-Id: If70c90ee9b1c0dba3cc3c1674345434aaf739251