Update patch set 1
Patch Set 1:

In my opinion, it is bad UX to reset QS tiles position when tuner changes so we could add it there.
I just didn't do it initially because i'm not sure what you guys think about that.

 > (1 comment)

In my opinion, it is bad UX to reset QS tiles position when tuner changes so we could add it there.
I just didn't do it initially because i'm not sure what you guys think about that.

Patch-set: 1