Update patch set 1
Patch Set 1:

> Doesn't look like he's going to respond.  I assume even a 2fps
 > improvement is worthwhile.

Hey, sorry for not replying, I've been busy changing teams so there's no one at Imagination who personally owns this anymore... but we're still *very* interested in getting this integrated!

As discussed with Chris over e-mail, we expect a significantly larger performance gain on lower-end platforms with less arithmetic performance, fugu's GPU is too high-end to benefit much. It has a 8:1 arithmetic to texture ratio, while our low-end cores have a 2:1 or 4:1 ratio, so will be much more limited by arithmetic operations which this optimises.

Based on Chris' testing and our own, the "if(dist.x+dist.y > 0.0)" conditional in ProgramCache.cpp is beneficial for us but may cause regressions on some platforms, so we agreed it'd make sense to integrate this change except for that part. Is that still the plan?

Cheers,
Arun

Patch-set: 1
Reviewer: Gerrit User 1136375 <1136375@85c56323-6fa9-3386-8a01-6480fb634889>