Update patch set 1
Patch Set 1:

I know this is a cherry-pick and we'll leave it of course, but just for the record:
I really don't like the commit message. It's devoid of any content and will loose all meaning to anyone but the author and including the author in a few weeks. We attempt to aim for messages that are as informative as is possible in "git log --oneline --no-merges". At least as informative as one can in one line....

Patch-set: 1
Reviewer: Gerrit User 1058798 <1058798@85c56323-6fa9-3386-8a01-6480fb634889>