IpSec Invalidate the Resource IDs on close()
To facilitate error checking we should invalidate
the resource IDs of all objects when we close() them.
Today, the resource ID is invalidated on the Transform
object but not on the SPI or Encap Socket. This CL
unifies the behavior.

In addition, checks for valid IP addresses are moved
from the IpSecConfig parcelable to the builder for
greater flexibility and consistency. The same checks
are already present in the IpSecService for security
reasons (but are not as user-friendly), which is why
we still want checks in both places.

Bug: 70641274
Test: cts - IpSecManagerTest
Change-Id: I28caec3e913902c748c6a50b4ef742ccef8b1b09