Can not get a thumbnail of Jpeg image of QCIF size.
Test: make WITH_TIDY=1 mmma art
Test: make test-art
Test: manual - took a photo
Test: refactoring CL. Existing unit tests still pass.

Bug Details:
Can not get a thumbnail of Jpeg image of QCIF (176 x 144) size.

Cause:
handleThumbnailFromJfif() has received the stream of content at ByteOrderedDataInputStream in, but since the read pointer is advanced in preprocessing,
the in.available() method can not acquire the expected value and calculates the wrong thumbnail Length.

When acquiring a thumbnail image of the content, in the case of a small size photograph such as QCIF,
the range check process of the content malfunctions and it becomes impossible to acquire the correct thumbnail size.

Modification:
Process of returning the pointer of the stream of the content to the head was added to normalize the range check processing.

Change-Id: I71fd49019edc628eced082334099058733963ee7
Signed-off-by: rie kanazawa <rie.kanazawa.sd@kyocera.jp>