Update patch set 2
Patch Set 2:

> (1 comment)
Because showWhenLocked may be used twice, and we don't want canShowWhenLocked() be called twice, so use this code style.

if (isTop || keyguardLocked) {
                showWhenLocked = !isInPinnedStack && r.canShowWhenLocked();
            }

We run monkey about 1 hour and calculated that before this change, containsShowWhenLockedWindow and containsDismissKeyguardWindow called  421240 times totally, and after this change it only called 10949 times.

When the system loading is heavy, the window's animate's thread will hold the mWindowMap lock continuous because it's high priority. and then AM's binder call will waiting this lock many times because of too many containsShowWhenLockedWindow and containsDismissKeyguardWindow calls.

So we believe this change can improve the performance about Activity start, and it will be more significant when the system load is heavy.

Patch-set: 2
Reviewer: Gerrit User 1084668 <1084668@85c56323-6fa9-3386-8a01-6480fb634889>