Framework: Add option to sampling-profile app startup
Add --profile-startup to the options of am start to use perfprofd to
profile startup of the given app, starting from just before the zygote
fork.

Sample invocation:

  adb shell am start --profile-startup -n com.android.gallery3d/.app.GalleryActivity

The result will be in the usual location for perfprofd, i.e,
/data/misc/perfprofd. Scripts in system/extras/perfprofd/scripts can be
used to analyze the resulting (compressed) proto.

Currently, profiling takes place for exactly three seconds. In a follow-up,
this may be restricted to the point that the app reports idle.

Note: perfprofd is only available on userdebug and eng builds.

Bug: 79266002
Test: m
Test: adb shell am start --profile-startup -n com.android.gallery3d/.app.GalleryActivity
Change-Id: I02ab219ea1ab88dfdc3452eca7de40e1ed71b09e