Update patch set 3
Patch Set 3: Code-Review+1

(3 comments)

Mostly LGTM but I we should have tests for this. Could you add tests to ConnectivityServiceTests that check that when we register a NetworkAgent, the network is created? It shouldn't be hard because we already have a mock for NetworkManagementService.

Patch-set: 3
Reviewer: Gerrit User 1000835 <1000835@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=+1