Update patch set 16
Patch Set 16:

Hi Lucas. I've added comments inline:

> I'm not sure if b/120007064 will really happen or not.
> According to the Vpn.java, it will bring "CONNECTING" state to register network.
> And ConnectivityService#updateNetworkInfo() will call
>     NetworkManagementService#createVirtualNetwork() immediately.

I've seen this happen, specifically for the test-network set of changes (which registers directly in CONNECTED state. (We can change that, and do connecting -> connected, but the race still exists).

> So, I don't think there will be a race condition for vpn network.
> Or do you mean another network which will set uid in NetworkCapabilities? (But I don't see other networks will do this)

The specific race is that the updating of UIDs is done after the network callbacks are notified, at which point we're racing against an app trying to bind to that Network.

Patch-set: 16
Reviewer: Gerrit User 1221725 <1221725@85c56323-6fa9-3386-8a01-6480fb634889>