Update patch set 31
Patch Set 31: Code-Review+1

> Patch Set 30:
> 
> Please check my comment in StatusBar.java.
> We want to align the camera key vibrate feedback as other key long press.
> There is different UX between lockscreen(or sleep) and interactive.
> 
> So far this patch works well without selinux.

Hey Eiichi, thank you for the pointers.
I've aligned the vibration pattern as you suggested. Also, vibrateForCameraGesture() is now private and gone from the IStatusBar* interfaces.

Pushed an update for sepolicy as well, I hope that one catches all the apps. It's a bit unfortunate that PhoneFallbackEventHandler is called/included from within the app and not the system server so we have to allow broad access.
Maybe there are views in Android that do not originate from an app, like during the setup wizard? We'll see with further testing and examination whether they need an `allow` as well.
I will be able to test on a real device next week, hopefully.

Patch-set: 31
Reviewer: Gerrit User 1421169 <1421169@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=+1