Update patch set 2
Patch Set 2:

> Patch Set 2: Code-Review-1
> 
> (1 comment)
> 
> Typically a Parcelable class stands on its own, and its the responsibility of the sending/receiving class to handle the null-ness; see note inline about how I expected this to implement Parcelable.

This was discussed on the original review, and I ended up discussing with you in person:
http://ag/c/platform/frameworks/base/+/4299924/1/core/java/android/util/TimestampedValue.java#92

TL;DR - the class is generic so and an instance is only parcelable if the value it holds is parcelable. So, I decided against making it part of the general contract / guarantee for the class; only under certain circumstances will an object of this type be parcelable. In those cases it is support via writeToParcel/readFromParcel (the docs on those methods and at the class level could be clearer to explain this).

Restricting the value to being parcelable by requiring the value type to be parcelable wouldn't work with Long / String, etc.

Patch-set: 2
Reviewer: Gerrit User 1038039 <1038039@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Presubmit-Ready=0