Rename VpnConfig.LEGACY_VPN to SETTING_VPN
This change renames the pseudo-package used for settings-based VPNs to
encompass Platform VPNs that are configured from the Settings VPNs menu.

App-based platform VPNs will use the package name of the app as a key,
while the settings-based ones will use the pseudo-package.

Once the Legacy VPN is removed (not just deprecated), the Settings app
should use the standard VpnManger APIs (with one or two tweaks to allow
for multiple profiles per app)

Bug: 146577177
Test: FrameworksNetTests passing
Change-Id: Id8039ffc6363959fcdb40a52548af23bd15dca02