Clear linger timers on non DDS networks
Linger Timers for a network are per Network Request. In scenarios where
WiFi replaces the default cellular network, cellular network on dds sub
is lingered for 30 sec on default Network Request. If there is a Voice
call on non-DDS Sub causing DDS to switch temporarily during
the 30 sec duration, cellular  network on new dds replaces the BG
Network Request and cellular network on old dds sub is
lingered for 5 sec on BG request.

During the linger timer update,  maximum value of all timers is selected
and a timer is started for it. So the network on older dds continues to
linger for more than 5 sec.

This fix handles the issue by clearing the timers on non dds network if
the existing timer is greater than the linger time set after DDS is
triggered.

Tests: DDS / Voice call related scenarios.
Change-Id: If5c352183db8862ee326ef29936e183b640d4f71