Clear calling identity before creating NetworkAgent
In the current design of Vpn.java, it will call
NetworkCapabilities#setUnderlyingNetworks() if underlyingNetworks
of VpnConfig is set by the VPN app, which means calling
NetworkCapabilities#setUnderlyingNetworks() is not the intended
behavior of the VPN app.
But in the constructor of NetworkAgent, it will check if the
caller has required permissions or not to set the underlying
network of NetworkCapabilities.
If the VPN app doesn't have required permissions, the
SecurityException will be thrown. Clear calling identity before
creating NetworkAgent can prevent this problem.

Bug: 205738644
Test: atest HostsideVpnTests
Change-Id: Iee2cc55c732fcdda75cd620e00ad66d97f3183c3