VolumeDialogImpl: Respect multiple visible rows in landscape
If the user receives a call or an alarm is ringing we are showing two
volume rows, the default (music) stream and the active stream, in this
example either alarm or ring. Currently the second row is not touchable
in landscape mode because the calculation of the touchable region does
not handle the case of multiple visible rows.

This patchset adds a check wether the ringer drawer extra size is bigger
than the additionally visible rows and then substracts the size of the
visible rows so that they stay touchable.

Test: While receiving a call in landscape both, the music row and the
      ring row, are touchable. Touches at the left of the volume panel
      are still going through the app behind the volume panel.
Change-Id: Icda80b7a98d5510747bc92ef12a3c7647131b18d