Revert "identity: add TEST_MAPPING file to run identity CTS in p..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (go/aca-get/e2f80155-b21a-4c8a-8b95-7b785bb7b515). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ic3b7fc1645ceb81a7d0ccbd1b0187e4e9a6b477b