AudioService: Fix LE Audio volume when placing a call
Recent changes to setLeAudioVolumeOnModeUpdate() function removed the
hardcoded parameter, which turned out to be a workaround. But the
comment about the workaround was in the function caller and not in the
function iself and was left unnoticed.

This change restores the workaround, but this time in the proper place,
where the comment about the workaround is placed.

Bug: 241501978
Tag: feature
Test: tested manually with BLE Audio headset
Change-Id: I1339eb94f574a4718a63d3dec826d785b4fef3ca