Revert "Add a method to note waking network packets to BatteryStats"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/7763c096-41c1-4141-984b-fdbb44bc7dd8). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I985a2759f8f20b2519d679901f7fc5c2dd19403a