Revert "Add a method to note waking network packets to BatteryStats"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/e74184d1-5104-48ee-9bbe-6454c29b91e2). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I9d486916a296836dbccc20daf797ca1a22c68293