Revert "Use toolkit instead of renderscript in SystemUI"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/eb5fe0e2-4f25-444d-9978-f66911bc7bca). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Id61a7fdccce69a04fb14284cb1ec86a0204b9d3e