Revert "Use toolkit instead of renderscript in SystemUI"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/4263a9b8-3ece-40ce-ab01-25229be3bf46). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ib7d088c4e4a72a0f2f7da09d44156704200fc04d