Revert "Update lint checks to their state on internal master"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/4cabd7bd-6cf3-4ff2-be15-67db91887690). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Idf023d55c7fca83325a3c07d6a8ea0186f1068d4