Revert "framework-nfc: Switch to "framework-module-defaults""
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/ba36f608-5353-44e9-ba38-87f992f86eb1). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Iae6c07bccf30abaead2e0cb73258b5499301139a