Revert "Update LocaleUtils.filterByLanguage due to addLikelySubt..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/c3b8af68-469c-4986-8344-fcf780cf1b35). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I4beb53eb53ba030be2c6f5b17441a9371507c164