Revert "Revert "Revert "Revert "Revert "Workaround build breakag..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/e6c4dd4a-92d4-4d61-b818-7cddf2435d6d). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 347840973
Change-Id: I68756de45075397b1cc784220953ecbecf05b141
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com