Revert "PointerIcon: Add leniency to hotspot validation for scal..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/0cf462ad-06d2-4f8e-8b65-dce1bea3775c). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I82a1b0b363147d9e0a0a5faad59b578ec8e29805
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com