Removing the call to LocalBluetoothManager and LocalBluetoothProfileManager
When pairing with a device, what we observed is the call to connectAllProfiles happens twice.
This is because LocalBluetoothManager is listening for changes when the bond state changes,
and is attempting to call connect device.

The VolumePanelDialog doesn't need to use LocalBluetoothManager to find the active A2DP device.
We can just ask the system manager for it.

Test: m
Flag: NONE
Change-Id: If0d894c5d6eef09c16f0f23731eb02625c4bd03a