Revert "Stop logging the AppIntegrityManager related logs (namel..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/29da478d-a178-4b1d-8f8b-9e6f262855a4). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 372330556
Change-Id: I2ee683301a2d0cf9290d659354595f394e194712
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com